new results with summary statistic approach (with and without bugfix)
I rewrote the inference script to allow for different inference methods and
implemented SPM's summary statistic approach. This runs much faster than ADVI. I
also now exclude trials in which the RT was very soon (100ms) after the onset of
the considered dot under the assumption that the MEG signal in these trials does
not contribute to the dot-specific effects we are after. The output is now not
only the negative log10 p-values, but also the underlying means and standard
deviations.

General Summary Statistics Results
----------------------------------
I only looked at the p-values so far, but these are promising and reach values
around 10^-8 for logpost_left. It is weird that in comparison to the ADVI
results the effects for surprise and logpost_left appear reversed, i.e., in the
summary statistic approach there is almost no effect of surprise (~10^-3), but a
reasonable effect of logpost_left. Perhaps this is related to the interaction
between the scaling of the regressors and the priors in ADVI and I should check
that.

I also observe strong effects of intercept and trial time with p-values smaller
than 10^-8 and moderate effects of dot_y and entropy (~10^-6).

The location of the effects seems to strongly overlap between logpost_left,
entropy and trial_time with perhaps small offsets in timing around 650ms. I
should investigate this.

dot_y appears to have two peaks, one at 600 ms and one at 700ms. Is this due to
the next dot being shown? But why is then no effect at 500 ms? Also, the
y-positions of the dots should be uncorrelated. So there should be no effect of
the ith dot at the (i+1)th time point. Perhaps this has something to do with the
other regressors in the sense that the dot_y effect only is apparent when the
effects of the other regressors are accounted for.

Effect of removed trials
------------------------
The notebook shows results for an old run in which the bug was still in effect,
i.e., only for subject 2 trials were removed, because of nans in the surprise
regressor. For these results I also kept trials with fast RTs. The results with
the new code after the bugfix are very similar, though.

Because of fewer number of trials you would expect p-values to increase, but
because you supposedly remove noise from the data you would also expect p-values
to decrease. The results are equally mixed. P-values changed slightly, but not
much. For dot_y, surprise and trial_time the minimum p-values increased while
they further decreased for the remaining regressors, e.g. logpost_left from
10^-8.57 to 10^-8.73. Except for logpost_left and surprise the mean p-values
across all times and channels tended to decrease.