major bugfix + normalisation + ADVI overhaul + save options
Bugfix
------
regressors in design matrix were not aligned with names in output, consequently
my interpretations and descriptions from last commit should be remapped (the
order of the regressors in the output followed that of r_names variable now
saved in the results such that the last topomap in the notebook always showed
the intercept and not the trial_time as the figure label suggested); see below
for re-interpretation of results

Normalisation
-------------
I now normalise the regressors and data across all subjects and time points to
equalise potential contributions and minimise interactions with the prior in the
Bayesian model. I remember wondering about a strange change in results for one
of the regressors with the summary statistics approach where I did not expect a
change, but this might have been because I remapped the regressors in the
results at the same time such that the change was actually in the intercept, as
one would expect.

ADVI overhaul
-------------
Made ADVI work now with the new results structure and save posterior mean and
std for the second-level parameters. Also now save neg log10 p-values as
originally intended instead of raw p-values.

Saving options
--------------
Made most options, including choices for prior, explicit and now save them
together with the results in the HDF5-file.

Results
-------
I have looked at t-values (directly related to neg-log10-p-values but signed),
mean parameter values and their standard deviation across subjects.

logpost_left and entropy reach the largest mlog10p values (both ~8.7) while
dot_y reaches 6.5 and surprise and intercept are around 5.5. The chosen time
points in the topomaps do not necessarily show these differences clearly,
though.

NOTE that the topomaps can be deceiving. For example, it looks like the
intercept reaches very high t-values around 8 which is supported by the reported
number during mouse hover over the region, but the high number appears to only
result through extrapolation of the channels. I tried to switch that off by
setting image_interp='nearest', but this did not affect the reported numbers.
This strange effect happens mostly at the edges of the shown head.

The following description of mlog10p time courses is based on selecting the
maximum values across all channels.

Entropy:
It's a bit strange that there is this large effect of entropy around 650 ms.
This should be unrelated to the 5th dot which I intended to investigate here,
because the entropy regressor is the same across all time points within the
trial. However, there could be an interaction with the 5th-dot-related
regressors. I should investigate whether the timing around 650 ms remains when I
use, e.g., 4th-dot regressors.

Intercept:
The mlog10p values of the intercept peak every 100 ms around 50 ms after dot
onset. This could reflect some general elevated activity related to processing
of individual dots. Additionally, the largest mlog10p-peak is around 400 ms
which is odd. Could this be a processing artefact?

Logpost_left:
mlog10p of the log-posterior for the 5th dot has 3 strong peaks. One at 425 ms,
a sharp one at 525 ms and the largest one around 575 ms. The early one is the
smallest but still reaches values around 6.5. I don't know how to interpret this
early peak. Maybe it's related to the log-posterior of the previously shown dot.

Dot_y:
The y-position of the dot appears to have an effect only from about 570-600 ms,
or so, but there is another strong effect (mlog10p ~6.5) at 700 ms. Why is
processing of the 5th-dot y-position so phasic?

Surprise:
supposedly is a direct measure of the x-position of the 5th-dot. It is the only
measure for which mlog10p values appear to steadily rise until reaching a peak
around 570 ms which is followed by a drop and another peak around 650 ms. The
peaks are only marginally above values of 5.

Trial_time:
Has no clear effects. There is a peak around 570 ms, but the baseline is
extremely low (2) and the peak is low, too (3.5).

I didn't look too closely at the mean parameter values, because they are
dominated by the intercept around 400 ms. Also the corresponding standard
deviations are dominated by the intercept which has about twice as high stds as
the other regressors. This indicates that the intercept absorbs a lot of
inter-individual effects.