normalised ADVI results and permutated summary statistics approach
ADVI results largely confirm summary statistics approach and permutations
indicate that large t-value peaks are reliable.

ADVI results
------------
When data and regressors are normalised, ADVI infers very similar second-level
parameters as the summary statistics approach, as can be seen in
MEG_hierarchical.ipynb. Specifically, the 'posterior t-values' which I compute
by simply dividing the mean of the posterior by its standard deviation show very
similar effects as the estimated t-values from the summary statitistics approach
except that ADVI posterior t-values have slightly lower magnitudes. This
verifies both ADVI and the summary statistics approach as suitable inference
under the given conditions.

I could now, however, change the assumptions in ADVI, e.g., allowing different
noise per subject. Also, it may make sense to revisit the subject-level
parameter posteriors in ADVI to investigate between-subject variance of effects.
Note that ADVI gives not only subject-level betas, but also subject-level
t-values through the posterior. Also, the hierarchical model could have
constrained the size of an effect in an individual based on the others, but this
depends on the priors and whether they simply allow to capture all
between-subject variance.

Permutations
------------
I exemplarily checked the reliability of inferred t-values by permuting the MEG
data.

First, I permuted simply across all data points, i.e., across subjects,
trials and time points, but not MEG channels. If I remember correctly, the
resulting mlog10p-values were all below 4 or 3.5, validating the peaks with
mlog10p-values above 4 that I found.

But then I thought that permuting across subjects and time points is not the
right thing to do and permuted only the trials within subjects. Further I used
the same permutation of trials across all subjects so that the random
correlations which exist between same trials across subjects are maintained in
the data. This should ensure that really only the across-trial effects that I
investigate with my regression are randomised. Note that the permutation was
different at every time point, but equal in each channel. Meaning that at least
time points (31) should give independent measures of the size of effects that
can be expected in random data.

I so far ran two of these permutations and the results are summarised in the
notebook in the form of time-courses of maximum log10p-values for all the
time points. As expected, the permutation of trials did not affect (much) the
estimates of the intercept; after all the overall mean of the data across trials
should have been unchanged. Except for very few single time points the maximum
mlog10p-values stayed below 4 and often below 3 across all regressors except the
intercept (expected as explained) and the surprise. For surprise, in the first
permutation run the highest peak is at the time of the highest peak resulting
from the unpermuted data (~590 ms) and they have similar height (~5.5). In the
second permutation run the highest peak is just after the second highest peak in
the unpermuted data (~660 ms vs ~680 ms) and has similar height (~5). It is
unclear to me why this occurs specifically for surprise. Could there be an
effect of surprise that does not depend on the relation of the regressor with
the particular trial data, but the correlations across subjects? Is this perhaps
related to surprise being long-tailed so that it picks up outliers in the data?