4th-dot regressors and smoothing vs. no smoothing
I have run two sanity checks to better understand the results.

Smoothing vs. no smoothing
--------------------------
As already mentioned in the last commit of helpers.py, whether or not I smooth
the data does not appear to affect the second-level results much. This is shown
in the corresponding section in MEG_hierarchical.ipynb where I plot the time
courses of maximum mlog10p-values across channels. It looks like for most
regressors the values are a bit lower without smoothing, but overall the
trajectories follow each other closely. Given the minimal differences it may be
best to not smooth the data as there is no clear motivation to do so.

4th-dot regressors
------------------
I ran my analysis using the 4th-dot regressors instead of the 5th-dot regressors
within the previously chosen 5th-dot time window. There were several
'significant' effects in the time courses of maximum mlog10p-values.

results that should be the same and surprise
............................................
trial_time and intercept have almost exactly the same results as for the 5th-dot
regressors. This is good, because they are exactly the same regressors in both
cases. It is worrying, though, that entropy, which is also the same regressor
across 4th and 5th dots, has a high peak around 680 ms in the 5th-dot
regression, but not in the 4th-dot regression. Instead, there is a very high
peak of 4th-dot surprise around 680 ms while there is only a moderate peak
around 660 ms in the 5th-dot regression. Intuitively this doesn't make much
sense: why should there be a surprise signal 300 ms after the dot? Note that
surprise and entropy are mildly anti-correlated (~-0.3) in all subjects for the
4th dot, but less so for the 5th dot (~-0.15). Otherwise, there doesn't appear
to be an effect of 4th-dot surprise at all.

dot_y
.....
There are strong effects of 4th-dot y-position at two time points (450-520 ms
and 620-700 ms) with a low dip in between them. The effects are therefore quite
different from those for the 5th-dot.

logpost_left
............
Other than logpost_left for the 5th dot, the measure has an overall elevated
effect within the time window instead of clear peaks. From about 470 ms the
maximum mlog10p-values stay in the range 4-7, mostly within 4-6. There is an
early peak at 440 ms which is shared between 4th and 5th dots. Don't know how to
interpret this.