including all dot regressors using dot_x instead of logpost
Given my insights about regressor correlations I have now switched from
logpost_left to dot_x for the regression allowing me to include the x-positions
of all dots up to the 5th as regressors into the GLM. At the same time, I have
added all dot_y and surprise regressors. The correlations between regressors in
this large GLM stay within 0.3 for all subjects as shown in
regressor_correlations.ipynb (next commit).

Results
-------
I looked at the dot-specific regressors by aligning them to dot onset, but first
I report the dot-independent regressors.

trial_time
..........
Still shows no reliable effects except that there appears to be a gradual
increase of the effect within a trial, but even at the end between 800 to 900 ms
the mlog10p-values stay below 3 and permutations regularly reach these values.
The effect is clearer, though, when the mean beta values are plotted, but they
also only reach values around 0.035 at the peak.

entropy
.......
Similar to trial_time entropy shows an increasing effect towards the end of a
trial with a peak at 830 or 840 ms (depending on whether you look at mean betas
or mlog10p). Compared to the values that the two permutations reach, the peak
appears to just be significant. Overall, the mean effect appears to be small
(peak at beta=0.06).

intercept
.........
shows roughly the same effects as before. In terms of mlog10p it has 3 peaks
above 5 at 120, 220 and 320 ms. These are followed by shallower peaks with
values around 3 at 450-460, 550-560, 650-670 and 760 ms. Permutations show
the same peaks with a bit larger deviations from about 780 ms. The mean betas
are very large in comparison to the other regressors. Even the small
mlog10p-peaks reach beta values of 0.15. It is interesting that the first large
peak in mlog10p at 120 ms is only a small peak in mean betas (value of 0.15).
This is followed by another same sized peak at 170 ms which is visible in
mlog10p, but looks just like a small dent. The peaks at 220 and 320 ms are also
extremely large in mean betas (values just above 0.25). Then it looks like the
next peak at 420 ms is interrupted and replaced by the peak at 450 ms which is
followed by the series of peaks in steps of 100 ms. These peaks also
correspond to a different effect across the scalp where the polarities between
the two groups of peaks seem to switched, similar to the effect of dot_y
described below.

Overall these results indicate that the intercept shows a mixture of two
effects. One which is probably related to pure processing of the dots and a
later one reflecting some other process which is also driven by the dots, but
doesn't show very clear effects anymore. As I'm looking at averages across
subjects this might have to do something with before and after the decision.
Perhaps many subjects make their decision after the 4th dot on average across
trials, i.e., before the 5th. Consequently, I should be able to see that the
time point at which the large 120 ms peaks vanish varies with the average
reaction time of subjects.

dot_x
.....
exhibits three very clear effects. First, there is a strong peak 120 ms after
dot onset. This peak is roughly the same hight for all dot regressors in
mlog10p. There is a bit more variance of the peak height in mean betas, but
there doesn't seem to be an order that corresponds to the order of the dots in
the trial. Beta values for that peak are around 0.1. The t-values show that
this effect is in the back of the head with large positivity above visual
cortex. This is the same effect I saw when looking at logpost_left.

The second effect is an equally large, or perhaps higher peak at 170 ms. The
mean betas suggest a higher peak (0.12) and mlog10p suggest a roughly equal
sized peak. Again variance in mean betas doesn't correspond to the order of the
dots (dots in order of peak height, highest first: 1, 5, 2, 4, 3). The effect is
focused in a central region perhaps towards parietal region. This could be what
O'Connel and Murphy call the centro-parietal positivity. Then, this would not be
necessarily related to the accumulated evidence, but to the momentary evidence.

The third effect is a sustained plateau starting around 300 ms and lasting for
at least 200 ms. My aligned plot doesn't show it, but the plateau vanishes after
200-300 ms which can be seen for dot 1, dot 2 and dot 3 across the full time
window. The underlying topology is also a central positivity which is perhaps a
bit more central than the peak at 170 ms.

Althought the mean betas are not as high as those for the intercept the
mlog10p-values are higher indicating that the dot_x correlations are more
consistent across subjects than the average effects captured by the intercept.

dot_y
.....
The y-position of the dots also shows strong effects, but which are more
variable in time than that for dot_x across the different dots. There are two
major peaks, one at 180 ms and one at 320 ms after dot onset. Especially the
first peak is quite dispersed across time, but is slightly stronger. Both peaks
reach values of 8 for mlog10p and 0.08 for mean betas, but not for all 5 dots.
The topology shows a vertically divided dipole along the midline with a
positivity on the left and a negativity on the right at 180 ms and switched
polarities at 320 ms. The location is midway between the back of the head and
the middle (parietal?). So this seems to be a nice effect, but it's unclear to
me what it shows.

Surprise
........
is a bit puzzling. There is no clear dot-onset-related effect across all 5 dots.
Perhaps there is a tiny, general elevation from about 150 ms, but without
consistent peaks. On the other hand, there appear to be relatively strong
effects of surprise at apparently random times, e.g., for dot 3 at 130 ms and
420 ms and for dot 2 at 620 ms. I wonder what these effects actually pick up as
there cannot be a surprise-effect of dot 3 at 130 ms. The effect at 420 ms for
dot 3 is 220 ms after its dot onset which could be true, but why is there
nothing for the other dots? Surprise is highly anti-correlated with the
log-likelihoods. Perhaps one of them is a better regressor, but why would it? Is
there something special about the suprise_3 regressor?