started with subject-level regression results + surprise variance
I went out to check subject-level regression results exemplarily for a specific
channel and time point (as suggested by second-level analysis). First, I wanted
to get a feeling for the effect of dot x-positions in the MEG data at the most
prominent times and locations. To do this I fitted the previous GLM, computed
predictions of the fitted GLM without the regressor of interest and then plotted
the residuals between data and predictions in relation to the left-out
regressor. This should show the correlation between the data and regressor that
is picked up the fitted GLM parameter value of the subsequently left-out
regressor.

Indeed, I can see this correlation in the plots, but not necessarily for all
subjects. An example is shown in the notebook for dot_x_4 at the late P300(?)
response at 700 ms and the channel with the largest effect (MEG0721). The good
news is that for all except 2 subjects the dot_x_4 regressor has a positive
beta value and the second_level result (mean of betas) is not driven by a few
outlying subjects.

I was a bit surprised by how weak the effect is when plotted like that. The
correlation between residuals and regressor is hard to spot by eye, because the
residuals range from about -700 to 700 (I did not normalise the data, because it
would have been a different normalisation than in my second_level analysis,
because I wouldn't normalise over time points) and the correlation is often
rather small, e.g., 0.2. I should not have been surprised by this, because the
mean betas on the second level were around 0.1 while here they are around 20. In
relation to the normalised data which should roughly range in [-3.5, 3.5] the
size of effect of 0.1 corresponds to the 20 in the non-normalised data. This
means that for one standard unit in dot x-position there is only an effect of
1/10th of a standard unit in the data while the data ranges across roughly 7
standard units.

Variability of surprise
-----------------------
At some point I realised that surprise is not only driven by dot x-positions,
but also by their y-positions. I verified that this is the case, i.e., that my
surprise regressors actually vary with y-positions when x-positions are roughly
fixed. I then realised that even for roughly fixed x-position and y-position
surprise actually varies in my regressors. One source of this variability is the
effect of the posterior beliefs from the last dot. To exclude also this effect I
looked at the flipped trials (trials 241 to 408 have exactly the same dot
positions as trials 1 to 168 except that their x-positions are sign-flipped) for
which surprise should be equal.

It turns out that surprise varies in the range of -0.2 to 0.2 standard units
even when the sequence of dot positions is controlled for. For the first dot
this is even larger in the range of -0.4 to 0.4. I believe that this variability
is due to my simulations with different samples of parameter values. Surprise is
affected by noisestd and prior. I try to remove variance due to noisestd and
variability of prior through averaging across simulations, but apparently the
average over 2000 simulations still produces this variability in surprise. That
surprise after the first dot varies more is probably because of the prior which
may have only an effect at the start of the trial.

I wonder whether it makes sense to include a surprise regressor which is only
based on the x-positions and ignores y-positions, as the subjects may also
ignore y-positions.