replacing suprise with absolute dot positions in GLM
Surprise appeared to not show reliable effects and I realised that subjects may
ignore the y-positions when computing surprise. So I created new surprise
measures which i) are momentary (only related to current dot) and ii) evaluate
surprise of the x- and y-coordinates of the dots separately. This lead to the
following new regressors:

abs_dot_x
abs_dot_y
momentary_surprise
momentary_surprise_x
momentary_surprise_y
momentary_surprise_pc1
momentary_surprise_pc2

The first two are the rawest measures of surprise: they simply tell how far from
0 the corresponding coordinate of the dot is. They are strongly related to
momentary_surprise_x and _y. The only difference being that abs_dot_x and _y are
linear with respect to dot coordinates while momentary surprise are parabolas.
At first I assumed that momentary_surprise_x would have a non-trivial form in
the area between the targets, but that's not the case. Momentary surprise
equally averages across the two generative models which results in surprise
values which are lowest at x=0. momentary_surprise_x and momentary_surprise_y
are not exactly equal functions of distance from 0, but they are both parabolas
which are very close to each other.

The surprise combined across those of the individual coordinates
(momentary_surprise) is strongly correlated to momentary_surprise_x and _y. This
is expected as momentary_surprise has to be large whenever one of the
single-coordinate surprises is large. The correlation is a bit larger to
momentary_surprise_x for the 4th dot, probably because of the particular samples
of dot positions in the experiment trials.

To get rid of the correlation between the combined and single-coordinate
surprises I projected the three surprises onto their 2 principal components (the
third component explains virtually no variance). The first component captures
the common variance across the combined and single-coordinate surprises while
the second component captures the difference between momentary_surprise_x and
_y. I just saw that for the 4th dot there is actually a mild correlation (0.2)
between the regressors corresponding to the projections onto these components.
This can happen, because I do the PCA across all dots.

I eventually decided to use the raw abs_dot_x and _y as regressors in the GLM.
This allows me to say that this first GLM uses (except for entropy) no
model-based measures, but only direct transformations of the stimulus. It also
gives a clear interpretation of the regressors and can still be related to
momentary surprise.

Second-level results
--------------------
The inclusion of the abs_dot regressors and exclusion of surprise left the large
effects of dot_x, _y and intercept virtually unchanged. At least, I couldn't so
far spot any clear differences. There is some evidence that including the
abs_dot regressors has removed a bit of an effect of dot_y around 100 ms after
dot onset, but this change is small (maximum mean betas reduced from ~0.055 to
~0.046 for dot_y_4).

I could also not see any differences for the (almost non-existing) effects of
trial_time and entropy.

The abs_dot regressors have only rather weak effects which tend not to be
consistent across dots. The clearest effect is at ~120 ms after dot onset.
Details follow:

abs_dot_x
.........
For the 2nd and 4th dots abs_dot_x shows a strong effect around 120 ms (maximum
mean betas of ~0.06). For the 1st dot this peak is slightly delayed (at 140 ms)
and lower (~0.047). For the 5th dot the peak at 120 ms is also there, but even
lower (~0.038). One could argue that the peak also exists for the 3rd dot, but
it's not actually higher than 3 other peaks one of which is 40 ms after dot
onset. The average across dots then shows a clear peak of moderate height
(0.047) at 120 ms with no other noteworthy peaks.

The topomap at 120 ms shows a dipole over the back of the head similar to that
seen for dot_y at 320 ms, but a bit rotated such that the dipole centre line
goes from back left to front right of the head.

abs_dot_y
.........
also exhibits a peak at 120 ms, but weaker and mostly driven by the 5th and 2nd
dots which reach maximum mean betas of ~0.05 then. The 3rd dot shows no effect
at 120 ms, but instead a large peak around 220 ms (beta ~0.063). It's unclear to
me why this happens. Why should abs_dot_y have a large effect particularly for
the 3rd dot at 220 ms instead of 120 ms? The effect is so large and extended
across time that it seems unlikely that it simply is due to noise. Similarly,
abs_dot_y shows a medium-sized (~0.05) but wide peak (330-400 ms) for the 4th
dot, which does not appear to be there for the other dots.

The t-value topomaps are equally confusing. The average effect at 120 ms is
pretty weak, but there appears to be a dipole over the back of the head which
has reversed polarity compared to that of abs_dot_x. The large peak for the 3rd
dot, on the other hand, corresponds to a dipole with the same polarity (negative
left) as that for abs_dot_x at 120 ms, but the orientation of the dipole centre
line is now tilted such that it goes from back right to top left. The large peak
for the 4th dot around 350 ms after its onset exhibits a more spread dipole
which otherwise looks quite similar to that of the dipole at 120 ms for
abs_dot_x. So all in all I don't currently know how to interpret these findings.

Conclusion
----------
I will stick with the abs_dot regressors. They may only show marginally
significant effects when looked at across dots, but I tend to believe that the
seen effects mean something. Also, the effect at 120 ms after dot onset appears
plausible and wasn't there for model-based surprise. Using abs_dot_x and _y also
made clear that the effect for the 3rd dot at 220 ms after its onset is more
related to the y-coordinate of the dot and not so much to the x-coordinate.
This couldn't be seen with model-based surprise.