investigating evidence accumulation specific regressor
Given that variance explained by dot_x will be removed from my following
analyses I thought about how I can capture specifically evidence
accumulation-related effects in a regressor. I decided to perform PCA on the
combination of dot_x and logpost_left, but I didn't like how logpost_left
appears to be heavy tailed (towards large negative values and bounded by 0). So
instead I introduced the log-posterior ratio (lpr) which is simply the DDM's
decision variable and is better behaved numerically.

PCA on dot_x and lpr produces two principal components. The first, explaining
over 70% of variance, is the sum of the two measures while the second is their
difference. So the second component is the one I'm interested in: it subtracts
the contribution of dot_x from lpr, supposedly leaving only the accumulated
evidence.

It turned out that it's not really possible to use a common PCA-derived
rotation to separate dot_x from accumulated evidence for all dots. For late
dots the accumulated evidence has more influence on the current lpr than dot_x
while its the reverse for early dots. That's why I decided to do the PCA only on
the first 5 dots.

Checking the correlation of the new regressor with dot_x I noticed that this
correlation varies strongly across subjects. While for most subjects this
(anti-)correlation is below 0.5, there are also subjects which still exhibit a
strong anti-correlation between dot_x and the potential accumulated evidence
regressor (highest: -0.82). When investigating how fitted parameter values
relate to this anti-correlation, I found that dtmode provides the best
explanation for the correlation (showing a strong 0.81 correlation between
itself and the size of the anti-correlation between dot_x and accumulated
evidence). This makes sense: if dtmode is small, e.g., below 0.4, evidence
cannot actually be accumulated - the subject is probably heavily driven by
momentary evidence. Then the accumulated evidence regressor is still very much
related to dot_x. On the other hand, if dtmode is large, the model explains a
lot of the subject responses by accumulating evidence from the dots. Then the
accumulated evidence regressor is actually less related to dot_x.

I originally saw this as motivation for doing the PCA on single-subject basis
which should account for the idiosyncratic evidence accumulation of subjects,
but realising that anyway the relative contribution of accumulated evidence in
lpr rises with the number of shown dots I wonder whether I should instead simply
choose the previous lpr value as accumulated evidence contribution for the
current dot.