subject-wise accumulated evidence regressor through PCA
Applying the PCA on lpr and dot_x within single subjects indeed reduces the
correlation between the resulting regressor (accev_pca) and dot_x. The
corresponding anti-correlation now stays below -0.5 (values closer to 0) for all
subjects.

As expected, accev_pca correlates strongly with lpr of the previous dot. So I
may also just use that. Interestingly, the correlation reduces for later dots:
0.91, 0.89, 0.85, 0.78 for dots 2, 3, 4 and 5, respectively. This may be an
effect of the increasing relative contribution of the accumulated evidence on
lpr for later dots which PCA cannot account for, because I did the PCA across
the first 5 dots.