insights into differences between surprise and momentary surprise
The overall offset between surprise and momentary surprise is due to the
different noisestds across subjects: larger noisestd leads to larger offset of
(averaged) surprise. Makes sense as more noise should also lead to more surprise
on average.

The difference between surprise and momentary surprise for the first dot was
mostly due to a bug in my computation of surprise in the discrete static gauss
model where I accidentally added the contribution of the prior twice not
realising that I was changing only a view of an array (see commit 0cef5c0
in rtmodels).

I have now recomputed surprise values. The overall picture is the same: there is
almost no effect of 'accumulated surprise'. The corresponding 2nd principal
component explains only around 5% of variance with a maximum of 7.5% for
subject 14. At least 'accumulated surprise' is pretty much uncorrelated with any
other regressor.