dot onset analysis looking explicitly for sequential effects
I thought that I could increase the power of my analyses by using the constraint
that I'm looking for consistent effects that repeat every 100 ms, i.e., with
every new dot. This was supposed to specifically help to identify correlates of
accumulated evidence which may be weak in comparison to the dot_x effect.
Originally I thought about running the analysis on the residuals of the
non-sequential analysis, but then decided to run a stand-alone analysis with all
regressors.

Concept
-------
At each time point I simply extend the data and regressors by the data and
regressors of the time points following in 100, 200, 300 and 400 ms. This means
that I model the sequence of 5 data points using the corresponding sequence of
regressors. For example, this looks for correlates of the particular sequence of
accumulated evidence in the data.

Compared to the single time point regression, this approach avoids the issue
that accumulated evidence of a dot will be strongly correlated with the
x-position of the next dot which can make regression results unreliable. For
example, if I included accumulated evidence for the first 5 dots in a single
time point regression together with their x-positions, I would always have
strong correlations between accev of dot i and dot_x of dot i+1. In the approach
here I only pair accev of dot i with dot_x of dot i which are uncorrelated.

I wondered first whether simply extending the data vector for the general linear
model, which is like adding new independent data points, actually implements the
constraint that the data should follow the particular sequence I want and not an
arbitrary order of the sequence. I confirmed that this should do what I want by
setting up a hidden markov model with certain transitions and deriving the
likelihood of the corresponding observations. This likelihood turned out to
simply be the product of the likelihoods of the individual observations
corresponding to what I do in the GLM.

Results
-------
mean vs. mlog10p
................
My impression from the results is that the mean is a superior measure over
mlog10p, or the t-value. It seems that the variance across subjects is rather
high and that this variance is badly estimated, because mlog10p appears to be
much more noisy than the mean parameter estimates. I mostly get this from the
results of accev and intercept where for accev I can interpret the mean
parameter estimates quite well while mlog10p shows something quite different and
less interpretable and for the intercept the mean parameter values show
beautifully regular peaks while mlog10p indicates non-significant results due to
large variance across subjects. I should check why the variance across subjects
appears so large and why the variance estimates appear so unreliable.

intercept
.........
Note that the intercept I have implemented in the sequential regression
essentially averages the effects at the 5 data points. You can see this by
comparing the previous, single time point intercept effects with the ones shown
here.

The additional averaging reveals previous dot-onset effects even more clearly:
There are two major evoked effects, one always 20 ms after dot onset and one
always 60 ms after dot onset. At the beginning of the trial the 20 ms effect
dominates, but gradually diminishes as the trial progresses. Simultaneously the
60 ms effect becomes gradually more important. This can also be seen in the
single time point analysis, but here the gradual shift from 20 to 60 ms is much
clearer.

It is not so easy to tell when exactly, e.g., the 20 ms effect follows the
presentation of a particular dot, because I average over 5 time points. Clearly,
there cannot be an effect exactly 20 ms after dot onset. This is confirmed by
the single time point regression where the intercept shows no specific effect
within the first 100 ms after first dot onset, but a small peak appears after
120 ms.

Anyway, what is clear is that at the beginning of the trial the 20 ms effect
dominates while at the end of the trial the 60 ms effect dominates. Could this
have something to do with the general states of deciding and having decided? The
topography of the effect shows that the 20 ms effect appears to be a mirrored
version of the 60 ms effect.

dot_x and dot_y
...............
dot_x shows pretty much the same effect as the average across the 5 dot_x
regressors in the single time point regression: a big peak at 120 ms followed by
another big peak at 170 ms followed by a trough followed by a high plateau from
about 300 ms.

Similarly, dot_y exhibits the same effects as before: a widish peak around 170
ms followed by a trough and a smaller peak at 320 ms.

Note that for the regressors which change with the dots I do not have trouble
locating the effects in time relative to individual dot onset, because the
sequence of dots is prescribed and I do not average across dots for these
regressors.

abs_dot_x and _y
................
Both show a small peak at 120 ms after dot onset, as also seen in the
corresponding single time point, dot onset aligned analysis.

accumulated evidence (accev)
............................
shows an effect whose strength is between that of dot_x and dot_y. The mean
betas indicate that accumulated evidence is continuously represented from as
early as 50 ms after dot onset with an intermediate dip around 180 ms and a
gradual decline after ~230 ms. In combination with the dot_x effects this
suggests that the initial processing of dot_x at 170 ms interferes with the
representation of accumulated evidence. Then the representation of accumulated
evidence is updated with the dot_x information until the new accumulated
evidence is represented, probably around 320 ms. I wonder, though, why the
accumulated evidence (of the previous dot) effect declines further from 350 to
500 ms, if the new evidence is already integrated around 320 ms.

That the accumulated evidence appears to be only properly represented at ~50 ms
is probably only because the first dot has no evidence at that time point yet.
The size of the effect is not 0, though, and is even in the range of abs_dot_x
from already 10 ms onwards. This could be a representation of a subject's bias
(the prior is the accumulated evidence for the first dot).

Note that mlog10p does not show the described effects. Instead, mlog10p suggests
that there is a reasonable effect from dot onset, but the effect increases
throughout the trial reaching a maximum at 370 ms after which it declines.

The topography of accev is relatively stable and resembles, unsurprisingly, that
of dot_x for the plateau from 300 ms. Remember that the signs of dot_x and accev
mean the opposite: negative values of dot_x support a decision towards left
while negative values of accev support a decision towards right. In the
committed results figures I have flipped the sign of the accev t-values to
correct for this discrepancy in interpretation.

entropy
.......
shows a very small effect towards the end of the considered period. The effect
is only as strong as that of accev at its lowest value at 500 ms.